requirements.txt
aiogram==3.13.1
groq==1.2.0
fastapi==0.115.4
uvicorn[standard]==0.32.0
sqlalchemy==2.0.36
aiosqlite==0.20.0
greenlet==3.1.1
jinja2==3.1.4
python-multipart==0.0.17
python-dotenv==1.0.1
itsdangerous==2.2.0
bcrypt==4.2.0
